Remove unused file. Change button icon
